requirements.txt
opencv-python
face_recognition
numpy
Pillow
torch
torchvision
progressbar2
